+++ R/dbandwidth.R
@@ -3,7 +3,7 @@ dbandwidth <-
            bwmethod = c("cv.cdf","normal-reference"),
            bwscaling = FALSE,
            bwtype = c("fixed","generalized_nn","adaptive_nn"),
-           ckertype = c("gaussian","truncated gaussian","epanechnikov","uniform"), 
+           ckertype = c("gaussian","epanechnikov","uniform"),
            ckerorder = c(2,4,6,8),
            ckerbound = c("none","range","fixed"),
            ckerlb = NULL,
@@ -52,9 +52,6 @@ dbandwidth <-
       bwtype = "fixed"
       ckertype = "gaussian"
     }
-
-    if (ckertype == "truncated gaussian" && ckerorder != 2)
-      .np_warning("using truncated gaussian of order 2, higher orders not yet implemented")
 
 
     ukertype = match.arg(ukertype)