+++ R/np.cdhat.helpers.R
@@ -330,8 +330,7 @@
     kerneval = switch(bws$ckertype,
       gaussian = CKER_GAUSS + bws$ckerorder / 2 - 1,
       epanechnikov = CKER_EPAN + bws$ckerorder / 2 - 1,
-      uniform = CKER_UNI,
-      "truncated gaussian" = CKER_TGAUSS
+      uniform = CKER_UNI
     ),
     ukerneval = switch(bws$ukertype,
       aitchisonaitken = UKER_AIT,