+++ R/np.condistribution.R
@@ -346,13 +346,13 @@ npcdist.condbandwidth <-
         xkerneval = switch(bws$cxkertype,
             gaussian = CKER_GAUSS + bws$cxkerorder/2 - 1,
             epanechnikov = CKER_EPAN + bws$cxkerorder/2 - 1,
-            uniform = CKER_UNI,
-            "truncated gaussian" = CKER_TGAUSS),
+            uniform = CKER_UNI
+),
         ykerneval = switch(bws$cykertype,
             gaussian = CKER_GAUSS + bws$cykerorder/2 - 1,
             epanechnikov = CKER_EPAN + bws$cykerorder/2 - 1,
-            uniform = CKER_UNI,
-            "truncated gaussian" = CKER_TGAUSS),
+            uniform = CKER_UNI
+),
         uxkerneval = switch(bws$uxkertype,
             aitchisonaitken = UKER_AIT,
             liracine = UKER_LR),