+++ R/np.density.R
@@ -182,8 +182,8 @@ npudens.bandwidth <-
     ckerneval = switch(bws$ckertype,
       gaussian = CKER_GAUSS + bws$ckerorder/2 - 1,
       epanechnikov = CKER_EPAN + bws$ckerorder/2 - 1,
-      uniform = CKER_UNI,
-      "truncated gaussian" = CKER_TGAUSS),
+      uniform = CKER_UNI
+),
     ukerneval = switch(bws$ukertype,
       aitchisonaitken = UKER_AIT,
       liracine = UKER_LR),