+++ R/np.plot.helpers.R
@@ -5345,8 +5345,7 @@
     kerneval = switch(state$bws$ckertype,
       gaussian = CKER_GAUSS + state$bws$ckerorder / 2 - 1,
       epanechnikov = CKER_EPAN + state$bws$ckerorder / 2 - 1,
-      uniform = CKER_UNI,
-      "truncated gaussian" = CKER_TGAUSS
+      uniform = CKER_UNI
     ),
     ukerneval = switch(state$bws$ukertype,
       aitchisonaitken = UKER_AIT,
@@ -9263,14 +9262,12 @@ plotFactor <- function(f, y, ...){
     xkerneval = switch(bws$cxkertype,
       gaussian = CKER_GAUSS + bws$cxkerorder / 2 - 1,
       epanechnikov = CKER_EPAN + bws$cxkerorder / 2 - 1,
-      uniform = CKER_UNI,
-      "truncated gaussian" = CKER_TGAUSS
+      uniform = CKER_UNI
     ),
     ykerneval = switch(bws$cykertype,
       gaussian = CKER_GAUSS + bws$cykerorder / 2 - 1,
       epanechnikov = CKER_EPAN + bws$cykerorder / 2 - 1,
-      uniform = CKER_UNI,
-      "truncated gaussian" = CKER_TGAUSS
+      uniform = CKER_UNI
     ),
     uxkerneval = switch(bws$uxkertype,
       aitchisonaitken = UKER_AIT,