+++ R/np.reghat.R
@@ -640,8 +640,7 @@ npreghat <-
   kernel.x.c <- switch(bws$ckertype,
     gaussian = CKER_GAUSS + bws$ckerorder / 2 - 1,
     epanechnikov = CKER_EPAN + bws$ckerorder / 2 - 1,
-    uniform = CKER_UNI,
-    "truncated gaussian" = CKER_TGAUSS
+    uniform = CKER_UNI
   )
   kernel.xu.c <- switch(bws$ukertype,
     aitchisonaitken = UKER_AIT,
@@ -775,8 +774,7 @@ npreghat <-
     kerneval = switch(bws$ckertype,
       gaussian = CKER_GAUSS + bws$ckerorder / 2 - 1,
       epanechnikov = CKER_EPAN + bws$ckerorder / 2 - 1,
-      uniform = CKER_UNI,
-      "truncated gaussian" = CKER_TGAUSS
+      uniform = CKER_UNI
     ),
     ukerneval = switch(bws$ukertype,
       aitchisonaitken = UKER_AIT,
@@ -1022,8 +1020,7 @@ npreghat <-
     kerneval = switch(bws$ckertype,
       gaussian = CKER_GAUSS + bws$ckerorder / 2 - 1,
       epanechnikov = CKER_EPAN + bws$ckerorder / 2 - 1,
-      uniform = CKER_UNI,
-      "truncated gaussian" = CKER_TGAUSS
+      uniform = CKER_UNI
     ),
     ukerneval = switch(bws$ukertype,
       aitchisonaitken = UKER_AIT,