+++ R/rbandwidth.R
@@ -7,7 +7,7 @@ rbandwidth <-
            bwmethod = c("cv.ls","cv.aic"),
            bwscaling = FALSE,
            bwtype = c("fixed","generalized_nn","adaptive_nn"),
-           ckertype = c("gaussian","truncated gaussian","epanechnikov","uniform"),
+           ckertype = c("gaussian","epanechnikov","uniform"),
            ckerorder = c(2,4,6,8),
            ckerbound = c("none","range","fixed"),
            ckerlb = NULL,
@@ -55,9 +55,6 @@ rbandwidth <-
     if (!any(kord == ckerorder))
       stop("ckerorder must be one of ", paste(kord,collapse=" "))
   }
-
-  if (ckertype == "truncated gaussian" && ckerorder != 2)
-    .np_warning("using truncated gaussian of order 2, higher orders not yet implemented")
 
   ukertype = match.arg(ukertype)
   okertype = match.arg(okertype)